Add category conjunction (AND, OR) for action views
The flexform plugin has now the possibility to select the category
conjunction. Besides the previous default OR, it is now possible to
explicitly select AND, OR and don't care.
Additionally, the DefaultEventConstraintsListener now handles all models
(incl. custom models) that use the default 'categories' column. Custom
overwrites can be removed.
Related #547